feat: Rename codejail dockerfile to codejail-service for consistency
The edx-internal argocd build scripts currently assume that the dockerfile
in this repo has the same name as the IDA's repo on GitHub. It doesn't
actually matter, here, since we're pulling from the repo directly rather
than ADDing the repo that the CD scripts download for us, but let's be
consistent.

Rename image as well.