feat: Adjust codejail-service apparmor profile for general use; add docs
Only functional change is to remove facility for terminal interaction
from inner profile; this turned out to not be very important for iterative
development. It was the only remaining dev-related aspect of the profile,
so now the big scary warning can be removed from the top (and replaced by
a different one).

This is preparatory for deploying to stage. Once we have a deployment
process, the docs here can be updated.