feat: Allow codejail-service webapp to receive kill signal (#114)
A lot of the new pods are in an attempting-termination status with hundreds
of the following `FailedKillPod` error event:

`unknown error after kill: runc did not terminate successfully: exit status 1: unable to signal init: permission denied`

I believe the issue is that the outer profile does not allow for receiving
signals. Hopefully this will fix it.