feat: update Credentials docker image to use Python 3.12
Updates the Credentials IDA's Dockerfile to install and use Python 3.12. We are also in the process of upgrading the Python version used by CI for the Credentials service too.

We also update the version of node to the latest version of Node 18 LTS (18.17.1 -> 18.20.7).

Included are minor updates and readability improvements so that this file is more in line with Dockerfiles for our other services.