feat: Make codejail Dockerfile build with no args; start linting
This commit adds CI checks for Dockerfiles, starting with the codejail
service. To make it build without errors and warnings, I had to remove the
string-based `CMD` directive. It was causing a `JSONArgsRecommended`
warning, but rather than switch to JSON args, I believe we should just
remove the CMD entirely -- we already override this in our devstack
compose file, and it's confusing to have this unused value.

BOMS-219