feat: Downgrade edxapp to Focal, with vendored Python
Revert commit 3f526a3 to get back to Focal, and then unbreak the build by using
vendored Python packages.

This use of vendored packages mirrors what we ended up doing on the EC2
pipeline. See BOMS-239 for background. We'll stay on Focal until
containerization of edxapp is complete, and then undo these changes and upgrade
to Jammy again.

Also remove explicit installation of `libpython3.11` and
`libpython3.11-stdlib` in main package list; these are dependencies that
will come in as part of installing python3.11/-dev, and shouldn't have
been included in the list.